+++ DartLang.PubSpec.Serialization.Yaml/Extensions/BuilderSkeletonExtensions.cs
@@ -3,11 +3,12 @@ using YamlDotNet.Serialization;
 
 namespace DartLang.PubSpec.Serialization.Yaml.Extensions;
 
-public static class SerializerSkeletonExtensions
+public static class BuilderSkeletonExtensions
 {
 	public static T WithPubSpecConverters<T>(this BuilderSkeleton<T> builder) where T : BuilderSkeleton<T>
 	{
 		return builder
+			.WithTypeConverter(UrlYamlConverter.Instance)
 			.WithTypeConverter(SemVersionYamlConverter.Instance)
 			.WithTypeConverter(DependencyMapYamlConverter.Instance)
 			.WithTypeConverter(PlatformsYamlConverter.Instance);